requirements.txt
# Data Processing
pandas==2.1.4
numpy==1.24.3

# API and Web
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0

# MinIO Client
minio==7.2.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Quality
great-expectations==0.18.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2

# Kafka (for later)
kafka-python==2.0.2

# LLM (for later)
langchain==0.0.352
transformers==4.36.2